Working tabs list
Added an extra <div> view in the HTML to show the working tabs. Created function to list working tabs in said <div> element. List auto-updates on click, and clicking an element in the list will remove it from the working tabs. Added required event listener. Refactored a bit of code to make it cleaner and more sensible. Added preliminary documentation for functions.