Profile functionality
Added basic profile functionality. Profiles store lists of tabs and can be switched through using the drop-down menu. Working tabs update on switch. Profiles can be deleted using the delete button, which asks for confirmation first. On backend, profiles are stored in profile_dict which is updated on start as well as on profile creation.