Wizard profile updating
Added profiling updating functionality to wizard. Pressing the submit button will now add all selected tabs to the selected profile, as well as update the display on both sides and save the changes to local storage.

Refactored listTabs to use replaceChildren structure so that it can be re-ran to update list of tabs once submit button is pressed.