Profile updating on events
Added functionality for updating profiles dynamically as tabs are edited within the window. All relevant events (creation, removal, moving, updating, and attaching/detaching) are linked to a function that reads the current tabs in the window and updates the currently active profile. This allows the user to keep their profiles up to date without having to re-save the profile every time they make a change.

Refactored wizard to use default profile instead of main window tabs, as well as play nice with the new background script functionality.